update: update theme zinc
