+++ package.json
@@ -9,9 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@nextui-org/react": "^2.2.10",
-    "@nextui-org/system": "^2.0.15",
-    "@nextui-org/theme": "^2.1.18",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
@@ -26,14 +23,14 @@
     "embla-carousel-autoplay": "^8.0.0",
     "embla-carousel-react": "^8.0.0",
     "framer-motion": "^11.0.20",
-    "lucide-react": "^0.358.0",
+    "lucide-react": "^0.363.0",
     "next": "14.1.3",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-dom": "^18",
     "sharp": "^0.33.2",
     "slugify": "^1.6.6",
-    "tailwind-merge": "^2.2.1",
+    "tailwind-merge": "^2.2.2",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
