+++ src/components/ui/film/film.card.tsx
@@ -25,12 +25,12 @@ export const FilmCard = (props: IMovieBase) => {
             src={thumb_url}
           />
           <div className="absolute inset-0 flex items-center justify-center gap-2 pointer-events-none opacity-0 group-hover:opacity-100 transition-opacity">
-            <PlayIcon className="w-16 h-16 scale-50 opacity-0 transition-all group-hover:scale-100 group-hover:opacity-100 bg-background bg-opacity-50 rounded-full p-4" />
+            <PlayIcon className="w-16 h-16 scale-50 opacity-0 transition-all group-hover:scale-100 group-hover:opacity-100 bg-accent bg-opacity-50 rounded-full p-4 text-primary" />
           </div>
         </div>
         <CardContent className="p-4">
           <h3 className="text-[15px] font-medium capitalize pt-1 block truncate line-clamp-1">{name}</h3>
-          <p className="text-sm truncate line-clamp-1 font-light">
+          <p className="text-sm truncate line-clamp-1 font-normal">
             {original_name}
           </p>
         </CardContent>
