+++ src/components/ui/film/film.carousel.tsx
@@ -23,7 +23,7 @@ export function FilmCarousel({ items = [], title }: TFilmCarouselProps) {
   return (
     <Fragment>
       {title ? (
-        <h2 className="text-md font-medium uppercase pl-2 border-l-[3px] border-solid border-tertiary">
+        <h2 className="text-md font-medium uppercase pl-2 border-l-[3px] border-solid border-primary text-primary">
           {title}
         </h2>
       ) : null}
