+++ src/providers/theme.tsx
@@ -3,12 +3,7 @@
 import * as React from "react";
 import { ThemeProvider as NextThemesProvider } from "next-themes";
 import { type ThemeProviderProps } from "next-themes/dist/types";
-import { NextUIProvider } from "@nextui-org/react";
 
 export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
-  return (
-    <NextUIProvider>
-      <NextThemesProvider {...props}>{children}</NextThemesProvider>
-    </NextUIProvider>
-  );
+  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
 }
