+++ tailwind.config.ts
@@ -1,16 +1,13 @@
-import { nextui } from "@nextui-org/react";
-import type { Config } from "tailwindcss";
-import { fontFamily } from "tailwindcss/defaultTheme";
+import type { Config } from "tailwindcss"
 
 const config = {
   darkMode: ["class"],
   content: [
-    "./pages/**/*.{ts,tsx}",
-    "./components/**/*.{ts,tsx}",
-    "./app/**/*.{ts,tsx}",
-    "./src/**/*.{ts,tsx}",
-    "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}", // for nextui
-  ],
+    './pages/**/*.{ts,tsx}',
+    './components/**/*.{ts,tsx}',
+    './app/**/*.{ts,tsx}',
+    './src/**/*.{ts,tsx}',
+	],
   prefix: "",
   theme: {
     container: {
@@ -21,9 +18,6 @@ const config = {
       },
     },
     extend: {
-      fontFamily: {
-        sans: ["var(--font-sans)", ...fontFamily.sans],
-      },
       colors: {
         border: "hsl(var(--border))",
         input: "hsl(var(--input))",
@@ -37,10 +31,6 @@ const config = {
         secondary: {
           DEFAULT: "hsl(var(--secondary))",
           foreground: "hsl(var(--secondary-foreground))",
-        },
-        tertiary: {
-          DEFAULT: "hsl(var(--tertiary))",
-          foreground: "hsl(var(--tertiary-foreground))",
         },
         destructive: {
           DEFAULT: "hsl(var(--destructive))",
@@ -84,74 +74,7 @@ const config = {
       },
     },
   },
-  plugins: [
-    require("tailwindcss-animate"),
-    nextui({
-      prefix: "nghien-phim",
-      layout: {
-        spacingUnit: 4, // in px
-        disabledOpacity: 0.5, // this value is applied as opacity-[value] when the component is disabled
-        dividerWeight: "1px", // h-divider the default height applied to the divider component
-        fontSize: {
-          tiny: "0.75rem", // text-tiny
-          small: "0.875rem", // text-small
-          medium: "1rem", // text-medium
-          large: "1.125rem", // text-large
-        },
-        lineHeight: {
-          tiny: "1rem", // text-tiny
-          small: "1.25rem", // text-small
-          medium: "1.5rem", // text-medium
-          large: "1.75rem", // text-large
-        },
-        radius: {
-          small: "8px", // rounded-small
-          medium: "12px", // rounded-medium
-          large: "14px", // rounded-large
-        },
-        borderWidth: {
-          small: "1px", // border-small
-          medium: "2px", // border-medium (default)
-          large: "3px", // border-large
-        },
-        
-      },
-      themes: {
-        light: {
-          layout: {
-            hoverOpacity: 0.8, //  this value is applied as opacity-[value] when the component is hovered
-            boxShadow: {
-              // shadow-small
-              small:
-                "0px 0px 5px 0px rgb(0 0 0 / 0.02), 0px 2px 10px 0px rgb(0 0 0 / 0.06), 0px 0px 1px 0px rgb(0 0 0 / 0.3)",
-              // shadow-medium
-              medium:
-                "0px 0px 15px 0px rgb(0 0 0 / 0.03), 0px 2px 30px 0px rgb(0 0 0 / 0.08), 0px 0px 1px 0px rgb(0 0 0 / 0.3)",
-              // shadow-large
-              large:
-                "0px 0px 30px 0px rgb(0 0 0 / 0.04), 0px 30px 60px 0px rgb(0 0 0 / 0.12), 0px 0px 1px 0px rgb(0 0 0 / 0.3)",
-            },
-          },
-        },
-        dark: {
-          layout: {
-            hoverOpacity: 0.9, //  this value is applied as opacity-[value] when the component is hovered
-            boxShadow: {
-              // shadow-small
-              small:
-                "0px 0px 5px 0px rgb(0 0 0 / 0.05), 0px 2px 10px 0px rgb(0 0 0 / 0.2), inset 0px 0px 1px 0px rgb(255 255 255 / 0.15)",
-              // shadow-medium
-              medium:
-                "0px 0px 15px 0px rgb(0 0 0 / 0.06), 0px 2px 30px 0px rgb(0 0 0 / 0.22), inset 0px 0px 1px 0px rgb(255 255 255 / 0.15)",
-              // shadow-large
-              large:
-                "0px 0px 30px 0px rgb(0 0 0 / 0.07), 0px 30px 60px 0px rgb(0 0 0 / 0.26), inset 0px 0px 1px 0px rgb(255 255 255 / 0.15)",
-            },
-          },
-        },
-      },
-    }),
-  ],
-} satisfies Config;
+  plugins: [require("tailwindcss-animate")],
+} satisfies Config
 
-export default config;
+export default config
